% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_my_data.R
\name{clean_my_data}
\alias{clean_my_data}
\title{Clean my FitBit data for classroom analyses}
\usage{
clean_my_data(my_fitbit_data)
}
\arguments{
\item{my_fitbit_data}{data structure saved from the `ECfitbitR::get_my_data(token=token)` function call.}
}
\description{
Clean my FitBit data for classroom analyses
}
