% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_heartrate.R
\name{collect_heartrate}
\alias{collect_heartrate}
\title{Collect fitbit heartrate data}
\usage{
collect_heartrate(token, start_date = NULL, period_to_collect = "1m",
  intraday = T)
}
\arguments{
\item{token}{The token you received from `fitbitr` function \code{\link[fitbitr]{oauth_token}}}

\item{start_date}{The date that you want to start collecting the data in "YYYY-MM-DD" format. If `NULL` (default) then it chooses yesterday.}

\item{period_to_collect}{Resolution of the data to collect. Default is `1m` (one month). Options are "1d", "7d", "30d", "1w", "1m".}

\item{intraday}{T/F for whether to collect intraday data}
}
\description{
Collect fitbit heartrate data
}
