% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_sleep.R
\name{collect_sleep}
\alias{collect_sleep}
\title{Collect fitbit sleep data}
\usage{
collect_sleep(token, start_date = NULL, days_prior_to_start = 30)
}
\arguments{
\item{token}{The token you received from `fitbitr` function \code{\link[fitbitr]{oauth_token}}}

\item{start_date}{The date that you want to start collecting the data in "YYYY-MM-DD" format. If `NULL` (default) then it chooses yesterday.}

\item{days_prior_to_start}{Days before `start_date` that you want to collect data from. Default is 30 days.}
}
\description{
Collect fitbit sleep data
}
\section{Citations}{

Much of the code in this function was learned from the excellent blog post https://www.r-bloggers.com/downloading-fitbit-data-histories-with-r/
}

