% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_my_data.R
\name{get_my_data}
\alias{get_my_data}
\title{Get my data}
\usage{
get_my_data(token, start_date = NULL)
}
\arguments{
\item{token}{The token you received from `fitbitr` function \code{\link[fitbitr]{oauth_token}}}

\item{start_date}{The date that you want to start collecting the data in "YYYY-MM-DD" format. If `NULL` (default) then it chooses yesterday.}
}
\description{
A wrapper function to collect all the data in a single step
}
